added maven plugin.

+++ sim-domain/src/main/java/com/sim/dao/InvoiceDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import com.sim.domain.Invoice;
 
-public interface InvoiceDAO extends DAO<Invoice, Long>{
+public interface InvoiceDAO extends DAO<Invoice, Integer>{
 
 	List<Invoice> getInvoicesByDate(Timestamp startDate, Timestamp endDate);
 }
